requirements.txt
pandas==2.2.2
watchdog==4.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tenacity==9.0.0
pytest==8.3.2